feat(client):console navigation and breadcrumbs
