+++ libs/ui/design-system/src/lib/components/Breadcrumbs/Breadcrumbs.scss
@@ -1,10 +1,14 @@
 @import "../../style/index.scss";
 
 .amp-breadcrumbs {
+  margin-left: var(--default-spacing);
+
   &__link {
     color: var(--gray-20);
     text-decoration: none !important;
     font-size: var(--default-font-size);
+    display: flex;
+    align-items: center;
 
     &:hover {
       text-decoration: none;
@@ -20,5 +24,13 @@
     color: var(--gray-20);
     margin-left: var(--icon-spacing);
     margin-right: var(--icon-spacing);
+  }
+
+  &__first-separator {
+    width: 1px;
+    height: var(--double-spacing);
+    background: var(--gray-70);
+    border: none;
+    margin-right: 0;
   }
 }
