+++ packages/amplication-client/src/Assistant/ConsoleNavigationButton.tsx
@@ -1,8 +1,5 @@
-import { EnumButtonStyle } from "@amplication/ui/design-system";
-import { Link } from "react-router-dom";
-import { Button } from "../Components/Button";
+import { EnumTabsStyle, Icon, Tabs } from "@amplication/ui/design-system";
 import { useAppContext } from "../context/appContext";
-import { AnalyticsEventNames } from "../util/analytics-events.types";
 import { useProjectBaseUrl } from "../util/useProjectBaseUrl";
 
 const CLASS_NAME = "console-navigation-button";
@@ -10,29 +7,33 @@ const CLASS_NAME = "console-navigation-button";
 const ConsoleNavigationButton = () => {
   const { currentProject } = useAppContext();
 
-  const { isPlatformConsole } = useProjectBaseUrl();
+  const { baseUrl: platformBaseUrl } = useProjectBaseUrl({
+    overrideIsPlatformConsole: true,
+  });
 
-  const { baseUrl } = useProjectBaseUrl({
-    overrideIsPlatformConsole: !isPlatformConsole,
+  const { baseUrl: catalogBaseUrl } = useProjectBaseUrl({
+    overrideIsPlatformConsole: false,
   });
 
   if (!currentProject) {
     return null;
   }
 
-  const label = !isPlatformConsole ? "Platform Console" : "Service Catalog";
-
   return (
-    <Link to={baseUrl} className={CLASS_NAME}>
-      <Button
-        buttonStyle={EnumButtonStyle.Text}
-        eventData={{
-          eventName: AnalyticsEventNames.AskJovuClick,
-        }}
-      >
-        {label}
-      </Button>
-    </Link>
+    <Tabs tabsStyle={EnumTabsStyle.Header} className={CLASS_NAME}>
+      <Tabs.Tab
+        name="Service Catalog"
+        to={catalogBaseUrl}
+        exact={false}
+        icon={<Icon icon="code" size="small" />}
+      />
+      <Tabs.Tab
+        name="Platform Console"
+        to={platformBaseUrl}
+        exact={false}
+        icon={<Icon icon="grid" size="small" />}
+      />
+    </Tabs>
   );
 };
