+++ packages/amplication-client/src/Platform/ProjectPlatformPage.tsx
@@ -2,12 +2,10 @@ import { EnumTextColor, TabItem } from "@amplication/ui/design-system";
 import React, { useContext, useMemo } from "react";
 import { match } from "react-router-dom";
 import PageLayout from "../Layout/PageLayout";
-import useBreadcrumbs from "../Layout/useBreadcrumbs";
 import useTabRoutes from "../Layout/useTabRoutes";
 import { AppContext } from "../context/appContext";
 import { AppRouteProps } from "../routes/routesUtil";
 import ServiceTemplateList from "./ServiceTemplateList";
-import { useProjectBaseUrl } from "../util/useProjectBaseUrl";
 
 type Props = AppRouteProps & {
   match: match<{
@@ -24,14 +22,8 @@ const ProjectPlatformPage: React.FC<Props> = ({
   tabRoutes,
   tabRoutesDef,
 }) => {
-  const { currentProject, pendingChanges } = useContext(AppContext);
+  const { pendingChanges } = useContext(AppContext);
 
-  const { baseUrl: projectUrl } = useProjectBaseUrl({
-    overrideIsPlatformConsole: false,
-  });
-
-  useBreadcrumbs(`${currentProject?.name} `, projectUrl);
-  useBreadcrumbs(`Platform Console`, match.url);
   const { tabs, currentRouteIsTab } = useTabRoutes(tabRoutesDef);
 
   //count how many unique resources in the pending changes
