+++ packages/amplication-client/src/Workspaces/WorkspacePage.tsx
@@ -1,12 +1,10 @@
-import React, { useContext } from "react";
+import { TabItem } from "@amplication/ui/design-system";
+import React from "react";
 import { match } from "react-router-dom";
 import PageLayout from "../Layout/PageLayout";
-import useBreadcrumbs from "../Layout/useBreadcrumbs";
 import useTabRoutes from "../Layout/useTabRoutes";
-import { AppContext } from "../context/appContext";
 import { AppRouteProps } from "../routes/routesUtil";
 import WorkspaceOverview from "./WorkspaceOverview";
-import { TabItem } from "@amplication/ui/design-system";
 
 type Props = AppRouteProps & {
   match: match<{
@@ -22,9 +20,6 @@ const WorkspacePage: React.FC<Props> = ({
   tabRoutes,
   tabRoutesDef,
 }) => {
-  const { currentWorkspace } = useContext(AppContext);
-  useBreadcrumbs(currentWorkspace?.name, match.url);
-
   const { tabs, currentRouteIsTab } = useTabRoutes(tabRoutesDef);
 
   const tabItems: TabItem[] = [
